Move "Properties" components into "Properties" folder
