+++ src/Microsoft.VisualStudio.ProjectSystem.CSharp.VS.UnitTests/Microsoft.VisualStudio.ProjectSystem.CSharp.VS.UnitTests.csproj
@@ -55,10 +55,10 @@
     <Compile Include="..\Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests\Properties\AssemblyAttributes.cs">
       <Link>Properties\AssemblyAttributes.cs</Link>
     </Compile>
-    <Compile Include="ProjectSystem\VS\CSharpProjectDesignerPageProviderTests.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\CSharpProjectDesignerPageProviderTests.cs" />
     <Compile Include="ProjectSystem\VS\Imaging\CSharpProjectImageProviderTests.cs" />
     <Compile Include="ProjectSystem\VS\LanguageServices\CSharpCodeDomProviderTests.cs" />
-    <Compile Include="ProjectSystem\VS\CSharpProjectGuidProviderTests.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\CSharpProjectGuidProviderTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests\App.config">
