+++ src/Microsoft.VisualStudio.ProjectSystem.VisualBasic.VS.UnitTests/Microsoft.VisualStudio.ProjectSystem.VisualBasic.VS.UnitTests.csproj
@@ -53,9 +53,9 @@
       <Link>Properties\AssemblyAttributes.cs</Link>
     </Compile>
     <Compile Include="ProjectSystem\VS\Imaging\VisualBasicProjectImageProviderTests.cs" />
-    <Compile Include="ProjectSystem\VS\VisualBasicProjectDesignerPageProviderTests.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\VisualBasicProjectDesignerPageProviderTests.cs" />
     <Compile Include="ProjectSystem\VS\LanguageServices\VisualBasicCodeDomProviderTests.cs" />
-    <Compile Include="ProjectSystem\VS\VisualBasicProjectGuidProviderTests.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\VisualBasicProjectGuidProviderTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
