+++ src/Microsoft.VisualStudio.ProjectSystem.VisualBasic.VS/Microsoft.VisualStudio.ProjectSystem.VisualBasic.VS.csproj
@@ -46,8 +46,8 @@
   <ItemGroup>
     <Compile Include="ProjectSystem\VS\LanguageServices\VisualBasicLanguageServiceHost.cs" />
     <Compile Include="ProjectSystem\VS\Imaging\VisualBasicProjectImageProvider.cs" />
-    <Compile Include="ProjectSystem\VS\VisualBasicProjectDesignerPage.cs" />
-    <Compile Include="ProjectSystem\VS\VisualBasicProjectDesignerPageProvider.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\VisualBasicProjectDesignerPage.cs" />
+    <Compile Include="ProjectSystem\VS\Properties\VisualBasicProjectDesignerPageProvider.cs" />
     <Compile Include="ProjectSystem\VS\LanguageServices\VisualBasicCodeDomProvider.cs">
       <SubType>Component</SubType>
     </Compile>
